+++ Libraries/LibWeb/Page/Page.h
@@ -16,13 +16,13 @@
 #include <AK/Weakable.h>
 #include <LibGC/Heap.h>
 #include <LibGC/Root.h>
+#include <LibGfx/Cursor.h>
 #include <LibGfx/Forward.h>
 #include <LibGfx/Palette.h>
 #include <LibGfx/Point.h>
 #include <LibGfx/Rect.h>
 #include <LibGfx/ShareableBitmap.h>
 #include <LibGfx/Size.h>
-#include <LibGfx/StandardCursor.h>
 #include <LibIPC/Forward.h>
 #include <LibURL/URL.h>
 #include <LibWeb/CSS/PreferredColorScheme.h>
